src/backend/pyproject.toml
[tool.poetry]
name = "training-app"
version = "0.1.0"
description = ""
authors = ["kolesnikovvitaliy <kolesnikovvitaliy@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
alembic = "^1.13.0"
pydantic = {extras = ["email"], version = "^2.5.2"}
pydantic-settings = "^2.1.0"
aioodbc = "^0.5.0"
python-dotenv = "^1.0.0"
setuptools = "^69.0.2"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
bcrypt = "4.0.1"



[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
